+++ pybamm/input/parameters/lithium_ion/negative_electrodes/graphite_Chen2020_composite/silicon_LGM50_electrolyte_exchange_current_density_Chen2020.py
@@ -1,11 +1,11 @@
 from pybamm import exp, constants
 
 
-def graphite_LGM50_electrolyte_exchange_current_density_Chen2020(
+def silicon_LGM50_electrolyte_exchange_current_density_Chen2020(
     c_e, c_s_surf, c_s_max, T
 ):
     """
-    Exchange-current density for Butler-Volmer reactions between graphite and LiPF6 in
+    Exchange-current density for Butler-Volmer reactions between silicon and LiPF6 in
     EC:DMC.
 
     References
@@ -32,10 +32,12 @@ def graphite_LGM50_electrolyte_exchange_current_density_Chen2020(
         Exchange-current density [A.m-2]
     """
 
-    m_ref = 6.48e-7  # (A/m2)(mol/m3)**1.5 - includes ref concentrations
+    m_ref = (
+        6.48e-7 * 28700 / 278000
+    )  # (A/m2)(mol/m3)**1.5 - includes ref concentrations
     E_r = 35000
     arrhenius = exp(E_r / constants.R * (1 / 298.15 - 1 / T))
 
     return (
-        m_ref * arrhenius * c_e**0.5 * c_s_surf**0.5 * (c_s_max - c_s_surf) ** 0.5
+        m_ref * arrhenius * c_e ** 0.5 * c_s_surf ** 0.5 * (c_s_max - c_s_surf) ** 0.5
     )