+++ pybamm/models/full_battery_models/base_battery_model.py
@@ -228,6 +228,7 @@ class BatteryModelOptions(pybamm.FuzzyDict):
                 "reaction-driven",
                 "stress and reaction-driven",
             ],
+            "open circuit potential": ["single", "current sigmoid"],
             "operating mode": [
                 "current",
                 "voltage",
@@ -519,6 +520,7 @@ class BatteryModelOptions(pybamm.FuzzyDict):
                                 "intercalation kinetics",
                                 "interface utilisation",
                                 "loss of active material",
+                                "open circuit potential",
                                 "particle mechanics",
                                 "particle",
                                 "particle phases",