+++ pybamm/parameters/parameter_sets.py
@@ -162,6 +162,18 @@ Chen2020 = {
     "citation": "Chen2020",
 }
 
+Chen2020_composite = {
+    "chemistry": "lithium_ion",
+    "cell": "LGM50_Chen2020",
+    "negative electrode": "graphite_Chen2020_composite",
+    "separator": "separator_Chen2020",
+    "positive electrode": "nmc_Chen2020",
+    "electrolyte": "lipf6_Nyman2008",
+    "experiment": "1C_discharge_from_full_Chen2020",
+    "sei": "example",
+    "citation": ["Chen2020", "Ai2022"],
+}
+
 Mohtat2020 = {
     "chemistry": "lithium_ion",
     "cell": "UMBL_Mohtat2020",